implementado calcu
